pyproject.toml
[project]
name = "seo-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "bs4>=0.0.2",
    "chromadb>=1.5.1",
    "dotenv>=0.9.9",
    "extruct>=0.18.0",
    "fastapi[standard]>=0.129.0",
    "faststream[rabbit]>=0.6.7",
    "greenlet>=3.3.1",
    "html-to-markdown>=2.24.5",
    "html2text>=2025.4.15",
    "langchain>=1.2.10",
    "langchain-core>=1.2.12",
    "langchain-openai>=1.1.9",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.8",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "matplotlib>=3.10.8",
    "mypy>=1.19.1",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "playwright>=1.58.0",
    "plotly>=6.5.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.12.1",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "pypdf>=6.1.3",
    "pytz>=2025.2",
    "pytesseract>=0.3.13",
    "retry-plus>=1.0.5",
    "safety>=3.7.0",
    "sentence-transformers>=5.2.3",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "types-passlib>=1.7.7.20260211",
    "types-python-dateutil>=2.9.0.20260305",
    "types-pytz>=2025.2.0.20251108",
    "ultimate-sitemap-parser>=1.8.0",
    "uvicorn>=0.40.0",
    "wordcloud>=1.9.6",
    "pillow>=12.0.0",
]
